Use ContiguousBytes protocol instead of Data
